function/requirements.txt
google-api-python-client>=2.0.2
google-cloud-storage>=1.37.0
tensorflow>=2.4.1
numpy>=1.19.5
